+++ src/Firebase/Exception/OutOfBoundsException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Kreait\Firebase\Exception;
+
+class OutOfBoundsException extends \OutOfBoundsException implements FirebaseException
+{
+}